Move tracking of blocks-with-tests to TabulatedResult.
